+++ app/ProfilePage/ActionItems.tsx
@@ -9,7 +9,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage'; // Assumin
 import { Cache } from 'react-native-cache';
 import { memoizedGetTheme } from 'helper/redux/settings';
 import { useSubscribe } from '@nostr-dev-kit/ndk-mobile';
-import { getLightningAmount } from 'components/cashu';
+import { getLightningAmount } from 'helper/cashuClient';
 
 const reactionCache = new Cache({
   namespace: 'reactions',
